Make Database a class again
Changing Database to an interface is a binary incompatible change.
Since we don’t want to change major versions, make it a class again for
now.
